replace int with size_t

+++ include/core/Algorithm.h
@@ -57,7 +57,7 @@ struct Algorithm {
 
   [[nodiscard]] Algorithm operator+(const Algorithm& other) const;
 
-  [[nodiscard]] Algorithm operator*(const int& times) const;
+  [[nodiscard]] Algorithm operator*(const size_t& times) const;
 
   [[nodiscard]] Algorithm withSetup(const std::string& setup_alg_string) const;
